requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
idna==2.8
importlib-metadata==0.23
jsonschema==3.2.0
more-itertools==7.2.0
packaging==19.1
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pyrsistent==0.15.5
pytest==5.1.2
pytest-cov==2.7.1
requests==2.25.0
six==1.12.0
termcolor==1.1.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.6.0
Flask==1.1.2
pymongo==3.11.1
jproperties==2.1.0
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
Flask-Jsonpify==1.5.0
Pillow==8.0.1
qrcode==6.1
gunicorn